put both types in a single file
