+++ lib/parachain/collation_protocol.go
@@ -1,6 +1,9 @@
 package parachain
 
-import "github.com/ChainSafe/gossamer/pkg/scale"
+import (
+	"github.com/ChainSafe/gossamer/lib/common"
+	"github.com/ChainSafe/gossamer/pkg/scale"
+)
 
 // CollationProtocol represents all network messages on the collation peer-set.
 type CollationProtocol scale.VaryingDataType
@@ -26,4 +29,70 @@ func (c *CollationProtocol) Set(val scale.VaryingDataTypeValue) (err error) {
 func (c *CollationProtocol) Value() (val scale.VaryingDataTypeValue, err error) {
 	vdt := scale.VaryingDataType(*c)
 	return vdt.Value()
+}
+
+// CollatorProtocolMessage represents Network messages used by the collator protocol subsystem
+type CollatorProtocolMessage scale.VaryingDataType
+
+// Index returns the index of varying data type
+func (CollatorProtocolMessage) Index() uint {
+	return 0
+}
+
+// NewCollatorProtocolMessage returns a new collator protocol message varying data type
+func NewCollatorProtocolMessage() CollatorProtocolMessage {
+	vdt := scale.MustNewVaryingDataType(Declare{}, AdvertiseCollation{}, CollationSeconded{})
+	return CollatorProtocolMessage(vdt)
+}
+
+// Set will set a value using the underlying  varying data type
+func (c *CollatorProtocolMessage) Set(val scale.VaryingDataTypeValue) (err error) {
+	vdt := scale.VaryingDataType(*c)
+	err = vdt.Set(val)
+	if err != nil {
+		return
+	}
+	*c = CollatorProtocolMessage(vdt)
+	return
+}
+
+// Value returns the value from the underlying varying data type
+func (c *CollatorProtocolMessage) Value() (val scale.VaryingDataTypeValue, err error) {
+	vdt := scale.VaryingDataType(*c)
+	return vdt.Value()
+}
+
+// Declare the intent to advertise collations under a collator ID, attaching a
+// signature of the `PeerId` of the node using the given collator ID key.
+type Declare struct {
+	CollatorId        CollatorID        `scale:"1"`
+	ParaID            uint32            `scale:"2"`
+	CollatorSignature CollatorSignature `scale:"3"`
+}
+
+// Index returns the index of varying data type
+func (Declare) Index() uint {
+	return 0
+}
+
+// AdvertiseCollation contains a relay parent hash and is used to advertise a collation to a validator.
+// This will only advertise a collation if there exists one for the given relay parent and the given peer is
+// set as validator for our para at the given relay parent.
+// It can only be sent once the peer has declared that they are a collator with given ID
+type AdvertiseCollation common.Hash
+
+// Index returns the index of varying data type
+func (AdvertiseCollation) Index() uint {
+	return 1
+}
+
+// CollationSeconded represents that a collation sent to a validator was seconded.
+type CollationSeconded struct {
+	Hash                         common.Hash                  `scale:"1"`
+	UncheckedSignedFullStatement UncheckedSignedFullStatement `scale:"2"`
+}
+
+// Index returns the index of varying data type
+func (CollationSeconded) Index() uint {
+	return 4
 }
